client/requirements.txt
paho-mqtt
pandas
requests
